requirements.txt
pytesseract
requests
pillow
fastapi
uvicorn[standard]